refactor pages to actually be pages and not components
